Extract NetworkProtocol to separate file
